+++ timefiller/_time_series_imputer.py
@@ -1,8 +1,9 @@
 import numpy as np
 import pandas as pd
-import statsmodels.api as sm
+from numba import njit, prange
 from sklearn.feature_selection import r_regression
-from sklearn.linear_model import Ridge
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import StandardScaler
 from tqdm.auto import tqdm
 
 from ._misc import check_params
@@ -10,85 +11,238 @@ from ._multivariate_imputer import ImputeMultiVariate
 
 
 class TimeSeriesImputer:
-    """Class for time series imputation.
+    """
+    Imputation of multivariate time series.
 
-    Args:
-        estimator (object, optional): Estimator used for imputation. Defaults to Ridge
-        preprocessing (callable, optional): Data preprocessing.
-        ar_lags (int, list, numpy.ndarray, or tuple, optional): Autoregressive lags to consider.
-        multivariate_lags (int or None, optional): Multivariate lags to consider.
-        na_frac_max (float, optional): Maximum fraction of missing values allowed.
-        min_samples_train (int, optional): Minimum number of samples for training.
-        weighting_func (callable, optional): Weighting function for imputation.
-        optimask_n_tries (int, optional): Number of attempts for optimization.
-        verbose (bool, optional): Display process details.
-        random_state (int or None, optional): Random state for reproducibility.
+    This class extends the capabilities of `ImputeMultiVariate` by accounting for
+    autoregressive and multivariate lags, as well as preprocessing.
     """
 
-    def __init__(self, estimator=None, preprocessing=None, ar_lags=None, multivariate_lags=None, na_frac_max=0.33,
-                 min_samples_train=50, weighting_func=None, optimask_n_tries=1, verbose=False, random_state=None):
-        if estimator is None:
-            estimator = Ridge(alpha=1e-5)
-        self.imputer = ImputeMultiVariate(estimator=estimator, preprocessing=preprocessing,
-                                          na_frac_max=na_frac_max, min_samples_train=min_samples_train,
-                                          weighting_func=weighting_func, optimask_n_tries=optimask_n_tries,
-                                          verbose=verbose)
+    def __init__(
+        self,
+        estimator=None,
+        alpha=None,
+        preprocessing=None,
+        ar_lags=None,
+        multivariate_lags='auto',
+        na_frac_max=0.33,
+        min_samples_train=50,
+        weighting_func=None,
+        optimask_n_tries=1,
+        negative_ar=False,
+        random_state=None,
+        verbose=0
+    ):
+        """
+        Initializes the imputer for time series.
+
+        Args:
+            estimator (object, optional): Estimator for regression.
+            alpha (float or list, optional): Quantile level(s) for prediction intervals.
+                Default is None.
+            preprocessing (object, optional): Preprocessing step (e.g., StandardScaler).
+                Default is StandardScaler(with_mean=False).
+            ar_lags (int, list, optional): Autoregressive lags to consider.
+                If int, generates lags from -ar_lags to +ar_lags (excluding 0).
+                Default is None.
+            multivariate_lags (int or str, optional): If specified, the optimal positive
+                and negative lags are determined for each column based on correlation.
+                These optimal lags are recalculated for each target column. If 'auto',
+                optimal lags are searched within 5% of the series length. If an integer,
+                optimal lags are searched within the range [-multivariate_lags, multivariate_lags].
+                Default is 'auto'.
+            na_frac_max (float, optional): Maximum fraction of missing values per column.
+                Default is 0.33.
+            min_samples_train (int, optional): Minimum number of training samples.
+                Default is 50.
+            weighting_func (callable, optional): Function to weight samples.
+                Default is None.
+            optimask_n_tries (int, optional): Number of attempts for OptiMask.
+                Default is 1.
+            negative_ar (bool, optional): If True, includes negative versions of
+                autoregressive lags. This can be useful when combined with a linear model
+                that has positive coefficients, as it enforces positive coefficients for
+                the covariates without sign constraints on the autoregressive features.
+                Default is False.
+            random_state (int, optional): Random seed. Default is None.
+            verbose (int or bool, optional): Verbosity level. Default is 0.
+        """
+        self.imputer = ImputeMultiVariate(
+            estimator=estimator,
+            alpha=alpha,
+            na_frac_max=na_frac_max,
+            min_samples_train=min_samples_train,
+            weighting_func=weighting_func,
+            optimask_n_tries=optimask_n_tries,
+            verbose=verbose
+        )
+        self.preprocessing = self._get_preprocessing(preprocessing)
         self.ar_lags = self._process_lags(ar_lags)
-        self.multivariate_lags = check_params(multivariate_lags, types=(int, type(None)))
-        self.verbose = bool(verbose)
+        self.multivariate_lags = check_params(multivariate_lags, types=(int, str, type(None)))
+        self.negative_ar = negative_ar
+        self.verbose = verbose
         self.random_state = random_state
 
     def __repr__(self):
-        params = ", ".join(f"{k}={getattr(self, k)}" for k in ('ar_lags', 'multivariate_lags'))
-        return f"TimeSeriesImputer({params})"
+        return f"TimeSeriesImputer(ar_lags={self.ar_lags}, multivariate_lags={self.multivariate_lags})"
+
+    @staticmethod
+    def _get_preprocessing(preprocessing):
+        """
+        Returns the preprocessing as a pipeline if necessary.
+
+        Args:
+            preprocessing (object, None, tuple, list): Preprocessing.
+
+        Returns:
+            object: Preprocessing or default pipeline.
+        """
+        if preprocessing is None:
+            return StandardScaler(with_mean=False)
+        elif isinstance(preprocessing, (tuple, list)):
+            return make_pipeline(preprocessing)
+        else:
+            return preprocessing
 
     @staticmethod
     def _process_lags(ar_lags):
-        check_params(ar_lags, types=(int, list, np.ndarray, tuple, type(None)))
-        if ar_lags is None:
-            return None
+        """
+        Processes autoregressive lags to ensure they are consistent (sorted list without 0).
+
+        Args:
+            ar_lags (int, list, tuple, np.ndarray, None): Autoregressive lags.
+
+        Returns:
+            list or None: List of lags or None if not specified.
+        """
         if isinstance(ar_lags, int):
-            ar_lags = list(range(-abs(ar_lags), 0)) + list(range(1, abs(ar_lags)+1))
-            return tuple(sorted(ar_lags))
-        if isinstance(ar_lags, (tuple, list, np.ndarray)):
-            ar_lags = [-k for k in ar_lags if k != 0] + [k for k in ar_lags if k != 0]
-            return tuple(sorted(set(ar_lags)))
+            return list(range(-abs(ar_lags), 0)) + list(range(1, abs(ar_lags)+1))
+        if isinstance(ar_lags, (list, tuple, np.ndarray)):
+            return sorted(set([k for k in ar_lags if k != 0]))
+        return None
 
     @staticmethod
     def _sample_features(data, col, n_nearest_features, rng):
+        """
+        Randomly selects the most relevant columns for imputation
+        based on correlation and data availability.
+
+        Args:
+            data (pd.DataFrame): Data.
+            col (str): Target column.
+            n_nearest_features (int): Number of features to select.
+            rng (np.random.Generator): Random number generator.
+
+        Returns:
+            list: List of selected columns.
+        """
         x = data.fillna(data.mean())
-        # computes pearson correlation between col and others series
         s1 = r_regression(X=x.drop(columns=col), y=x[col])
-        # computes the mean number of timestamps containing common valid data between col and others series
-        s2 = ((~data[col].isnull()).astype(float).values@(~data.drop(columns=col).isnull()).astype(float).values)/len(data)
-        # features are sampled according those computed features
+        s2 = ((~data[col].isnull()).astype(float).values @ (~data.drop(columns=col).isnull()).astype(float).values) / len(data)
         p = np.sqrt(abs(s1) * s2)
         size = min(n_nearest_features, len(s1), len(p[p > 0]))
         cols_to_sample = list(data.drop(columns=col).columns)
         return list(rng.choice(a=cols_to_sample, size=size, p=p/p.sum(), replace=False))
 
     @staticmethod
-    def _best_lag(s1, s2, max_lags):
-        c1 = sm.tsa.ccf(s1, s2, nlags=max_lags)[::-1]
-        c2 = sm.tsa.ccf(s2, s1, nlags=max_lags)[1:]
-        c = np.concatenate([c1, c2])
-        return np.abs(c).argmax() - max_lags + 1
+    @njit(parallel=True, boundscheck=False)
+    def cross_correlation(s1, s2, max_lags):
+        """
+        Computes cross-correlation between two series with lags.
+        Equivalent to [pd.Series(s1).corr(pd.Series(s2.shift(lag))) for lag in range(-max_lags, max_lags+1)],
+        but much faster: NaNs are taken into account, and the function uses Numba plus the Welford algorithm.
+        Args:
+            s1 (array): First series.
+            s2 (array): Second series.
+            max_lags (int): Maximum lag to compute.
+
+        Returns:
+            tuple: Lags and cross-correlation values.
+        """
+        n = len(s1)
+        cross_corr = np.zeros(2 * max_lags + 1, dtype=np.float32)
+        lags = np.arange(-max_lags, max_lags+1)
+        for k in prange(len(lags)):
+            lag = lags[k]
+            m1, m2 = 0., 0.
+            v1, v2 = 0., 0.
+            count = 0
+            cov = 0.
+            for i in range(n):
+                j = i + lag
+                s1i, s2j = s1[i], s2[j]
+                if (j >= 0) and (j < n) and np.isfinite(s1i) and np.isfinite(s2j):
+                    m1u = (count*m1 + s1i)/(count+1)
+                    m2u = (count*m2 + s2j)/(count+1)
+                    if count != 0:
+                        v1 = ((count - 1)*v1 + (s1i-m1)*(s1i-m1u))/(count)
+                        v2 = ((count - 1)*v2 + (s2j-m2)*(s2j-m2u))/(count)
+                    if count != 0:
+                        cov += (s1i - m1u)*(s2j-m2)/count
+                        cov *= (count/(1+count))
+                    count += 1
+                    m1, m2 = m1u, m2u
+            cross_corr[lag+max_lags] = count/(count-1)*cov/np.sqrt(v1*v2)
+        return lags, cross_corr
+
+    @classmethod
+    def _best_multivariate_lag(cls, s1, s2, max_lags):
+        """
+        Finds the best multivariate lags for a series `s2` relative to `s1`.
+
+        Args:
+            s1 (pd.Series): Reference series.
+            s2 (pd.Series): Series to lag.
+            max_lags (int or str): Maximum lags or 'auto'.
+
+        Returns:
+            list: List of lags offering the best correlation.
+        """
+        assert len(s1) == len(s2)
+        if max_lags == 'auto':
+            max_lags = int(0.05 * len(s1))
+        lags, cc = cls.cross_correlation(s1=s1.values, s2=s2.values, max_lags=max_lags)
+        ret, cc0 = [0], cc[max_lags]
+        if cc[lags > 0].max() >= cc0:
+            ret.append(lags[lags > 0][cc[lags > 0].argmax()])
+        if cc[lags < 0].max() >= cc0:
+            ret.append(lags[lags < 0][cc[lags < 0].argmax()])
+        return ret
 
     @classmethod
     def find_best_lags(cls, x, col, max_lags):
-        df = x.fillna(x.mean())
-        cols = df.drop(columns=col).columns
+        """
+        Finds and applies the best multivariate lags to a given column.
+
+        Args:
+            x (pd.DataFrame): Data.
+            col (str): Target column.
+            max_lags (int or str): Maximum lags or 'auto'.
+
+        Returns:
+            pd.DataFrame: DataFrame with lagged series.
+        """
+        cols = x.drop(columns=col).columns
         ret = [x[col]]
         for other_col in cols:
-            lag = cls._best_lag(df[col], df[other_col], max_lags=max_lags)
-            if lag != 0:
+            ret.append(x[other_col])
+            lags = cls._best_multivariate_lag(x[col], x[other_col], max_lags=max_lags)
+            for lag in lags:
                 ret.append(x[other_col].shift(-lag).rename(f"{other_col}{-lag:+d}"))
-            else:
-                ret.append(x[other_col])
         return pd.concat(ret, axis=1)
 
     @staticmethod
     def _process_subset_cols(X, subset_cols):
+        """
+        Transforms the subset of columns into indices.
+
+        Args:
+            X (pd.DataFrame): Data.
+            subset_cols (None, str, list, tuple): Subset of columns.
+
+        Returns:
+            list: Indices of the columns.
+        """
         _, n = X.shape
         columns = list(X.columns)
         if subset_cols is None:
@@ -100,10 +254,21 @@ class TimeSeriesImputer:
                 return []
         if isinstance(subset_cols, (list, tuple, pd.core.indexes.base.Index)):
             return [columns.index(_) for _ in subset_cols if _ in columns]
-        raise TypeError(f"subset_cols should be of type str, list, tuple, or pandas Index. Received type {type(subset_cols)} instead.")
+        raise TypeError()
 
     @staticmethod
     def _process_subset_rows(X, before, after):
+        """
+        Selects a subset of rows based on `before` and `after` dates.
+
+        Args:
+            X (pd.DataFrame): Data indexed by time.
+            before (str, optional): Upper date limit.
+            after (str, optional): Lower date limit.
+
+        Returns:
+            list: Indices of the retained rows.
+        """
         index = pd.Series(np.arange(len(X)), index=X.index)
         if before is not None:
             index = index[index.index <= pd.to_datetime(str(before))]
@@ -112,50 +277,136 @@ class TimeSeriesImputer:
         return list(index.values)
 
     def _impute_col(self, x, col, subset_rows):
-        if isinstance(self.multivariate_lags, int):
-            x = self.find_best_lags(x, col, self.multivariate_lags)
-        x = x.copy()
-        if self.ar_lags is not None:
-            for k in sorted(self.ar_lags):
-                x[f"{col}{k:+d}"] = x[col].shift(k).copy()
-        index_col = list(x.columns).index(col)
-        x_col_imputed = self.imputer(x.values, subset_rows=subset_rows, subset_cols=index_col)[:, index_col]
-        return pd.Series(x_col_imputed, name=col, index=x.index)
-
-    def __call__(self, X, subset_cols=None, before=None, after=None, n_nearest_features=None) -> pd.DataFrame:
-        """Call method for imputation.
+        """
+        Imputes a specific column in a DataFrame.
 
         Args:
-            X (DataFrame): Data to be imputed. Constant features (i.e., features with zero standard deviation) will not be imputed nor used for the imputation of other series.
-            subset_cols (str, list, tuple, or pandas.core.indexes.base.Index, optional): Columns to be imputed. By default, all columns will be imputed, except constant features.
-            before (str or pd.Timestamp or None, optional): Date before which the data is imputed. By default, no lower temporal limit is set.
-            after (str or pd.Timestamp or None, optional): Date after which the data is imputed. By default, no upper temporal limit is set.
-            n_nearest_features (int, optional): Number of nearest features to consider. A heuristic is used: the features are selected randomly, based on
-            their correlations with the feature to be imputed, as well as the number of common temporal observations with the feature to be imputed.
-            Constant features are excluded from this selection.
+            x (pd.DataFrame): Data (including other columns potentially used
+                for imputation).
+            col (str): Target column to impute.
+            subset_rows (list): Rows to consider.
 
         Returns:
-            DataFrame: Imputed data.
+            pd.Series or (pd.Series, pd.DataFrame): Imputed series and optionally a
+            DataFrame of uncertainties if alpha is defined.
         """
+        if isinstance(self.multivariate_lags, int) or self.multivariate_lags == 'auto':
+            x = self.find_best_lags(x, col, self.multivariate_lags)
+        if self.ar_lags is not None:
+            x_ar = []
+            for k in sorted(self.ar_lags):
+                x_ar.append(x[col].shift(k).rename(f"{col}{k:+d}"))
+                if self.negative_ar:
+                    x_ar.append(-x[col].shift(k).rename(f"{col}{k:+d}_neg"))
+            x = pd.concat([x, pd.concat(x_ar, axis=1)], axis=1)
+        index_col = x.columns.get_loc(col)
+        if self.imputer.alpha is None:
+            x_col_imputed = self.imputer(x.values, subset_rows=subset_rows, subset_cols=index_col)[:, index_col]
+            return pd.Series(x_col_imputed, name=col, index=x.index)
+        else:
+            x_imputed, uncertainties = self.imputer(x.values, subset_rows=subset_rows, subset_cols=index_col)
+            x_imputed_col = x_imputed[:, index_col]
+            uncertainties_col = uncertainties[:, :, index_col]
+            alphas = sorted(sum([[a/2, 1-a/2] for a in self.imputer.alpha], []))
+            return (pd.Series(x_imputed_col, name=col, index=x.index),
+                    pd.concat([pd.Series(_, index=x.index, name=alpha) for _, alpha in zip(uncertainties_col, alphas)], axis=1))
 
-        rng = np.random.default_rng(self.random_state)
-        X_ = check_params(X, types=pd.DataFrame).copy()
-        check_params(X_.index, types=pd.DatetimeIndex)
+    def _preprocess_data(self, X):
+        """
+        Applies preprocessing to the data and sets a time frequency.
 
+        Args:
+            X (pd.DataFrame): Original data.
+
+        Returns:
+            pd.DataFrame: Preprocessed data.
+        """
+        if self.preprocessing is not None:
+            X_ = pd.DataFrame(self.preprocessing.fit_transform(X), index=X.index, columns=X.columns)
+        else:
+            X_ = X.copy()
         if X_.index.freq is None:
             X_ = X_.asfreq(pd.infer_freq(X_.index))
         X_ = X_[X_.columns[X_.std() > 0]].copy()
-        columns = list(X_.columns)
+        return X_
 
+    def _select_imputation_features(self, X_, col, n_nearest_features, rng):
+        """
+        Selects the most relevant features for imputing a column.
+
+        Args:
+            X_ (pd.DataFrame): Preprocessed data.
+            col (str): Target column.
+            n_nearest_features (int or None): Number of most relevant features
+                to select.
+            rng (np.random.Generator): Random number generator.
+
+        Returns:
+            list: List of selected column names.
+        """
+        if isinstance(n_nearest_features, int):
+            return [col] + self._sample_features(X_, col, n_nearest_features, rng)
+        else:
+            return list(X_.columns)
+
+    def __call__(self, X, subset_cols=None, before=None, after=None, n_nearest_features=None) -> pd.DataFrame:
+        """
+        Imputes missing values in a time series DataFrame.
+
+        Args:
+            X (pd.DataFrame): Data to impute, indexed by time.
+            subset_cols (None, str, list, optional): Columns to impute. Default is None
+                for all columns.
+            before (str, optional): Upper date limit for row selection.
+            after (str, optional): Lower date limit for row selection.
+            n_nearest_features (int, optional): Number of most relevant features
+                to select. The selection is randomized : covariates highly-correlated are
+                more likely to be selected ; covariates wich are not likely to be available when the
+                imputed columns is are not likely selected. If None, uses all available features.
+
+        Returns:
+            pd.DataFrame or (pd.DataFrame, dict): Imputed data and, if alpha is
+            defined, a dictionary containing uncertainties for each column.
+
+        Note:
+            About n_nearest_features, the probability of selecting a feature is proportional to:
+            sqrt(|correlation| * co_occurrence_ratio)
+            where:
+            - correlation is calculated using r_regression between the feature and target
+            - co_occurrence_ratio is the proportion of rows where both feature and target are non-null
+        """
+        rng = np.random.default_rng(self.random_state)
+        check_params(X, types=pd.DataFrame)
+        check_params(X.index, types=pd.DatetimeIndex)
+
+        X_ = self._preprocess_data(X)
+
+        columns = list(X_.columns)
         ret = [pd.Series(index=X.index)]
+        uncertainties = {}
+
         subset_rows = self._process_subset_rows(X_, before, after)
         subset_cols = self._process_subset_cols(X_, subset_cols)
+
         for index_col in tqdm(subset_cols, disable=(not self.verbose)):
             col = columns[index_col]
-            if isinstance(n_nearest_features, int):
-                cols_in = [col] + self._sample_features(X_, col, n_nearest_features, rng)
-            else:
-                cols_in = list(X_.columns)
-            ret.append(self._impute_col(x=X_[cols_in], col=col, subset_rows=subset_rows))
+            if X_[col].isnull().mean() > 0:
+                cols_in = self._select_imputation_features(X_, col, n_nearest_features, rng)
+                if self.imputer.alpha is None:
+                    ret.append(self._impute_col(x=X_[cols_in], col=col, subset_rows=subset_rows))
+                else:
+                    imputed_col, uncertainties_col = self._impute_col(x=X_[cols_in], col=col, subset_rows=subset_rows)
+                    ret.append(imputed_col)
+                    uncertainties[col] = uncertainties_col
         ret = pd.concat(ret, axis=1).reindex_like(X).combine_first(X)
-        return ret
+
+        if self.preprocessing is not None:
+            ret = pd.DataFrame(self.preprocessing.inverse_transform(ret), columns=ret.columns, index=ret.index)
+
+        if self.imputer.alpha is None:
+            return ret
+        else:
+            for col in uncertainties:
+                df = uncertainties[col]
+                uncertainties[col] = pd.DataFrame(self.preprocessing.inverse_transform(df), columns=df.columns, index=df.index)
+            return ret, uncertainties