+++ docs/source/whatsnew.rst
@@ -3,6 +3,11 @@
 What's New ?
 ============
 
+Version 1.0.10 (December 23th 2024)
+---------------------------------
+- Improved and faster handling of lag discovery for covariates
+- ``PositiveOutout`` has a `column` argument to specify the columns to treat as positive
+
 Version 1.0.9 (December 6th 2024)
 ---------------------------------
 - Major speed-up, sub-arrays creation handled by numba